+++ Core/Interfaces/ICreateBareGit.cs
@@ -1,4 +1,4 @@
-﻿namespace MigrateFromSvnToGit {
+﻿namespace MigrateFromSvnToGit.Interfaces {
     public interface ICreateBareGit {
         void Create(string projectName);
     }
