+++ Core/Interfaces/IProcessCaller.cs
@@ -1,4 +1,4 @@
-﻿namespace MigrateFromSvnToGit {
+﻿namespace MigrateFromSvnToGit.Interfaces {
     public interface IProcessCaller {
         void Execute(string fileName, string arguments);
     }
